@@ -1,15 +1,17 @@
 [package]
 name = "interpol-rs"
-version = "0.1.0"
-edition = "2018"
+version = "0.2.1"
+edition = "2021"
 
 [build-dependencies]
 cbindgen = "0.20.0"
 
 [dependencies]
+derive_builder = "0.10.2"
 lazy_static = "1.4.0"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
+typetag = "0.1.8"
 
 [lib]
 crate-type = ["cdylib"] # shared library (.so)
