@@ -1,6 +1,6 @@
 #pragma once
 
-/* Warning, this file is autogenerated by cbindgen. Don't modify this manually. */
+// Warning, this file is autogenerated by cbindgen. Don't modify this manually.
 
 #include <stdarg.h>
 #include <stdbool.h>
@@ -9,128 +9,106 @@
 #include <stdlib.h>
 
 
+typedef int32_t MpiRank;
+
+typedef uint64_t Tsc;
+
+typedef double Usecs;
+
 typedef int32_t MpiComm;
 
-typedef uint32_t MpiReq;
+typedef int32_t MpiTag;
 
-#define WORLD 0
+typedef int32_t MpiReq;
 
 /**
- * Registers an `MPI_Init` call into the static `TRACES` vector.
+ * Registers an `MPI_Init` call into a static vector.
  */
-void register_init(uint64_t cycles, double time);
+void register_init(MpiRank current_rank, Tsc tsc, Usecs time);
 
 /**
- * Registers an `MPI_Send` call into the static `TRACES` vector.
+ * Registers an `MPI_Init_thread` call into a static vector.
  */
-void register_send(uint64_t cycles_lo,
-                   uint64_t cycles_hi,
-                   size_t bytes,
-                   MpiComm comm,
-                   int32_t current_rank,
-                   int32_t partner_rank,
-                   int32_t tag);
+void register_init_thread(MpiRank current_rank,
+                          Tsc tsc,
+                          Usecs time,
+                          int32_t required_thread_lvl,
+                          int32_t provided_thread_lvl);
 
 /**
- * Registers an `MPI_Recv` call into the static `TRACES` vector.
- */
-void register_recv(uint64_t cycles_lo,
-                   uint64_t cycles_hi,
-                   size_t bytes,
-                   MpiComm comm,
-                   int32_t current_rank,
-                   int32_t partner_rank,
-                   int32_t tag);
-
-/**
- * Registers an `MPI_Isend` call into the static `TRACES` vector.
- */
-void register_isend(uint64_t cycles_lo,
-                    uint64_t cycles_hi,
-                    size_t bytes,
-                    MpiComm comm,
-                    MpiReq req,
-                    int32_t current_rank,
-                    int32_t partner_rank,
-                    int32_t tag);
-
-/**
- * Registers an `MPI_Irecv` call into the static `TRACES` vector.
- */
-void register_irecv(uint64_t cycles_lo,
-                    uint64_t cycles_hi,
-                    size_t bytes,
-                    MpiComm comm,
-                    MpiReq req,
-                    int32_t current_rank,
-                    int32_t partner_rank,
-                    int32_t tag);
-
-/**
- * Registers an `MPI_Bcast` call into the static `TRACES` vector.
- */
-void register_bcast(uint64_t cycles_lo,
-                    uint64_t cycles_hi,
-                    size_t bytes,
-                    MpiComm comm,
-                    int32_t current_rank,
-                    int32_t partner_rank);
-
-/**
- * Registers an `MPI_Ibcast` call into the static `TRACES` vector.
- */
-void register_ibcast(uint64_t cycles_lo,
-                     uint64_t cycles_hi,
-                     size_t bytes,
-                     MpiComm comm,
-                     MpiReq req,
-                     int32_t current_rank,
-                     int32_t partner_rank);
-
-/**
- * Registers an `MPI_Gather` call into the static `TRACES` vector.
- */
-void register_gather(uint64_t cycles_lo,
-                     uint64_t cycles_hi,
-                     size_t bytes_s,
-                     size_t bytes_r,
-                     MpiComm comm,
-                     int32_t current_rank,
-                     int32_t partner_rank);
-
-/**
- * Registers an `MPI_Igather` call into the static `TRACES` vector.
- */
-void register_igather(uint64_t cycles_lo,
-                      uint64_t cycles_hi,
-                      size_t bytes_s,
-                      size_t bytes_r,
-                      MpiComm comm,
-                      MpiReq req,
-                      int32_t current_rank,
-                      int32_t partner_rank);
-
-/**
- * Registers an `MPI_Wait` call into the static `TRACES` vector.
- */
-void register_wait(uint64_t cycles_lo,
-                   uint64_t cycles_hi,
-                   MpiReq req,
-                   int32_t current_rank);
-
-/**
- * Registers an `MPI_Barrier` call into the static `TRACES` vector.
- */
-void register_barrier(uint64_t cycles_lo,
-                      uint64_t cycles_hi,
-                      MpiComm comm,
-                      int32_t current_rank);
-
-/**
- * Registers an `MPI_Finalize` call into the static `TRACES` vector.
+ * Registers an `MPI_Finalize` call into a static vector.
  *
- * As this *should* be the final registered event, serializes the contents of the `TRACES` vector.
+ * As this *should* be the final registered event, the contents of the vector will be sorted with
+ * every other MPI processes vectors' and then serialized.
  */
-void register_finalize(uint64_t cycles,
-                       double time,
-                       int32_t rank);
+void register_finalize(MpiRank current_rank,
+                       Tsc tsc,
+                       Usecs time);
+
+/**
+ * Registers an `MPI_Send` call into a static vector.
+ */
+void register_send(MpiRank current_rank,
+                   MpiRank partner_rank,
+                   uint32_t nb_bytes,
+                   MpiComm comm,
+                   MpiTag tag,
+                   Tsc tsc,
+                   Tsc duration);
+
+/**
+ * Registers an `MPI_Recv` call into a static vector.
+ */
+void register_recv(MpiRank current_rank,
+                   MpiRank partner_rank,
+                   uint32_t nb_bytes,
+                   MpiComm comm,
+                   MpiTag tag,
+                   Tsc tsc,
+                   Tsc duration);
+
+/**
+ * Registers an `MPI_Isend` call into a static vector.
+ */
+void register_isend(MpiRank current_rank,
+                    MpiRank partner_rank,
+                    uint32_t nb_bytes,
+                    MpiComm comm,
+                    MpiReq req,
+                    MpiTag tag,
+                    Tsc tsc,
+                    Tsc duration);
+
+/**
+ * Registers an `MPI_Irecv` call into a static vector.
+ */
+void register_irecv(MpiRank current_rank,
+                    MpiRank partner_rank,
+                    uint32_t nb_bytes,
+                    MpiComm comm,
+                    MpiReq req,
+                    MpiTag tag,
+                    Tsc tsc,
+                    Tsc duration);
+
+/**
+ * Registers an `MPI_Barrier` call into a static vector.
+ */
+void register_barrier(MpiRank current_rank,
+                      MpiComm comm,
+                      Tsc tsc,
+                      Tsc duration);
+
+/**
+ * Registers an `MPI_Test` call into a static vector.
+ */
+void register_test(MpiRank current_rank,
+                   MpiReq req,
+                   bool finished,
+                   Tsc tsc,
+                   Tsc duration);
+
+/**
+ * Registers an `MPI_Wait` call into a static vector.
+ */
+void register_wait(MpiRank current_rank, MpiReq req, Tsc tsc, Tsc duration);
