@@ -1,7 +1,5 @@
-pub mod event;
+#![feature(try_reserve_kind)]
+
 pub mod interpol;
-
-type MpiComm = i32;
-type MpiReq = u32;
-
-pub const WORLD: MpiComm = 0;
+pub mod mpi_events;
+pub mod types;
